Crate session

Modules

context

hints

session_config

table_name

Structs

MutableInner ๐

A container for mutable items in query context

Session

Session for persistent connection such as MySQL, PostgreSQL etc.

Enums

ReadPreference

Defines the read preference for frontend route operations, determining whether to read from the region leader or follower.

Type Aliases

SessionRef